refact: split to method composition class
Add some readability
