+++ src/plugins/Workers/SyncWorker/utils/TransactionOrderer/methods/reset.js
@@ -1,0 +1,7 @@
+/**
+ * Reset the local arrays
+ */
+module.exports = function reset() {
+  this.transactions = [];
+  this.transactionIds = [];
+};
